+++ src/collective/classification/folder/content/indexers.py
@@ -1,16 +1,18 @@
 # -*- coding: utf-8 -*-
 
-from Acquisition._Acquisition import aq_parent
-from collective.classification.folder.content.classification_folder import (
-    IClassificationFolder,
-)
-from collective.classification.folder.content.classification_subfolder import (
-    IClassificationSubfolder,
-)
-from collective.dexteritytextindexer import IDynamicTextIndexExtender
+from Acquisition._Acquisition import aq_parent  # noqa
+from collective.classification.folder.content.classification_folder import IClassificationFolder
+from collective.classification.folder.content.classification_subfolder import IClassificationSubfolder
+from collective.dexteritytextindexer.interfaces import IDynamicTextIndexExtender
 from plone.indexer.decorator import indexer
 from zope.component import adapter
 from zope.interface import implementer
+
+
+@indexer(IClassificationFolder)
+def archived_classification_folder_index(obj):
+    """Indexer for the archived field of a classification folder using yesno_value index. """
+    return obj.archived and '1' or '0'
 
 
 @indexer(IClassificationFolder)