+++ CHANGES.rst
@@ -4,8 +4,8 @@ Changelog
 1.0.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Corrected tests.
+  [sgeulette]
 
 1.0.2 (2023-11-28)
 ------------------