+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-1.0.3 (unreleased)
+1.1.0 (2024-02-19)
 ------------------
 
 - Corrected tests.